Added thin wrapper of 2D and 3D matrix calculations
Both 3D and 2D functionality is available in a single wrapped function
(available in both sequential and parallel versions). Of course, this
wrapping comes with a theoretical penalty, but the wrapped version
allows for a more flexible workflow. e.g. if the dimension of the
points transitions from 2D to 3D (or the converse) in the course of an
iterative process. Rather than handling dimension checking in the
external script, this library is able to handle it.

	new file:   benchmark/benchmark_wrapper.py
	modified:   src/distm/__init__.py
	modified:   src/distm/distm.py
	new file:   tests/test_wrapper.py